Reverted back to normal
Yuvan if you're seeing this, the plan moving forward is for you to bring your computer to school so that we can run the test on Nathan and possibly others. We don't have enough time to do the webrtc and change it to user end, since it is extremely complicated and not necessary.